+++ include/swgLib/pgrf.hpp
@@ -41,7 +41,7 @@ namespace ml
 		std::size_t read(std::istream& file);
 
 	protected:
-		uint8_t _version; // 0000, 0001
+		uint32_t _version; // 0000, 0001
 
 		// Path Node data...
 		std::vector<int32_t> _index;
